添加了参数调整，修改了一点车牌与字符分割的获取

+++ src/classifier/PlateRecognition_V3.cpp
@@ -142,7 +142,7 @@ PlateInfo PlateRecognition_V3::GetPlateInfo(PlateInfo &plateInfo,
 
     auto combineVerticalOrigin = [](vector<CharInfo> &charInfos,
                                     PlateInfo &plateInfo) {
-        for (size_t index = 1; index < min(charInfos.size(), 3ul); ++index) {
+        for (size_t index = 1; index < min((unsigned long)charInfos.size(), 3ul); ++index) {
             CharInfo &lastCharInfo = charInfos[index - 1];
             Rect &lastRect = charInfos[index - 1].OriginalRect;
             Rect &rect = charInfos[index].OriginalRect;
